Modificación de vouchers con referencia a los clientes + productos

+++ TPIII/Dominio/Voucher.cs
@@ -11,8 +11,8 @@ namespace Dominio
         public Int64 ID { get; set; }
         public string CodigoVoucher { get; set; }
         public bool Estado { get; set; }
-        public Int32 ? IdCliente { get; set; }
-        public Int32 ? IdProducto { get; set; }
+        public Int64 ? IdCliente { get; set; }
+        public Int64 ? IdProducto { get; set; }
         public DateTime ? FechaRegistro { get; set; }
 
         public Voucher(Int64 _ID, string _CodigoVoucher, bool _Estado, Int32 _IdCliente, Int32 _IdProducto, DateTime _FechaRegistro)
